app/requirements.txt
django>=4.0.0,<4.1.0
pyyaml
requests
django-cors-headers
Pillow
mysqlclient
crispy-bootstrap5==0.6
django-crispy-forms==1.14.0